+++ src/main/java/io/kontur/eventapi/dao/mapper/FeedMapper.java
@@ -67,4 +67,6 @@ public interface FeedMapper {
                       @Param("eventDetails") String eventDetails,
                       @Param("enriched") Boolean enriched,
                       @Param("episodes") String episodes);
+
+    Integer getNotEnrichedEventsCount();
 }