+++ src/test/java/io/kontur/eventapi/job/EnrichmentJobTest.java
@@ -20,6 +20,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import static io.kontur.eventapi.enrichment.EnrichmentConfig.*;
 import static org.mockito.Mockito.*;
@@ -60,8 +61,8 @@ public class EnrichmentJobTest {
         when(feedDao.getNotEnrichedEventsForFeed(feedWithEnrichment.getFeedId())).thenReturn(List.of(notEnrichedFeedData));
         doNothing().when(feedDao).addAnalytics(any());
         when(konturAppsClient.graphql(isA(InsightsApiRequest.class))).thenReturn(createResponse());
-        EventEnrichmentTask enrichmentTask = new EventEnrichmentTask(konturAppsClient, feedDao, postProcessors);
-        EnrichmentJob enrichmentJob = new EnrichmentJob(new SimpleMeterRegistry(), feedDao, enrichmentTask);
+        EventEnrichmentTask enrichmentTask = new EventEnrichmentTask(konturAppsClient, feedDao, postProcessors, new AtomicInteger(0), new AtomicInteger(0));
+        EnrichmentJob enrichmentJob = new EnrichmentJob(new SimpleMeterRegistry(), feedDao, enrichmentTask, new AtomicInteger(0), new AtomicInteger(0), new AtomicInteger(0));
 
         enrichmentJob.run();
 
@@ -79,8 +80,8 @@ public class EnrichmentJobTest {
         when(feedDao.getNotEnrichedEventsForFeed(feedWithEnrichment.getFeedId())).thenReturn(List.of(notEnrichedFeedData));
         doNothing().when(feedDao).addAnalytics(any());
         when(konturAppsClient.graphql(isA(InsightsApiRequest.class))).thenReturn(createErrorResponse());
-        EventEnrichmentTask enrichmentTask = new EventEnrichmentTask(konturAppsClient, feedDao, postProcessors);
-        EnrichmentJob enrichmentJob = new EnrichmentJob(new SimpleMeterRegistry(), feedDao, enrichmentTask);
+        EventEnrichmentTask enrichmentTask = new EventEnrichmentTask(konturAppsClient, feedDao, postProcessors, new AtomicInteger(0), new AtomicInteger(0));
+        EnrichmentJob enrichmentJob = new EnrichmentJob(new SimpleMeterRegistry(), feedDao, enrichmentTask, new AtomicInteger(0), new AtomicInteger(0), new AtomicInteger(0));
 
         enrichmentJob.run();